requirements.txt
requests
requests-toolbelt
kafka-python